+++ examples/avr/at90can/vscp_user/vscp_config_overwrite.h
@@ -126,6 +126,8 @@ extern "C"
 
 #define VSCP_CONFIG_DM_ROWS                     10
 
+#define VSCP_CONFIG_DM_NG_PAGE                  2
+
 #define VSCP_CONFIG_DM_NG_RULE_SET_SIZE         80
 
 #define VSCP_CONFIG_LOOPBACK_STORAGE_NUM        4